+++ src/crypto/common/mod.rs
@@ -1,7 +1,7 @@
 pub mod key_derivation {
-    use crate::crypto::cipher_suite::CipherSuite;
+    use crate::crypto::cipher_suite::CipherSuiteParams;
 
-    pub fn expand_subsecret(cipher_suite: &CipherSuite, key: &[u8]) -> (Vec<u8>, Vec<u8>) {
+    pub fn expand_subsecret(cipher_suite: &CipherSuiteParams, key: &[u8]) -> (Vec<u8>, Vec<u8>) {
         let aes_keysize = key.len() - cipher_suite.hash_len;
         let enc_key = key[..aes_keysize].to_vec();
         let auth_key = key[aes_keysize..].to_vec();