+++ src/frame/media_frame.rs
@@ -109,7 +109,7 @@ impl<'ibuf> MediaFrameView<'ibuf> {
     ) -> Result<EncryptedFrameView<'obuf>> {
         let key_id = key.key_id();
         log::trace!(
-            "Encrypting MediaFrame # {} using KeyId {} and CipherSuite {}",
+            "Encrypting MediaFrame # {} using KeyId {} and CipherSuiteParams {}",
             self.counter,
             key_id,
             key.cipher_suite_variant()