+++ src/ratchet/ratcheting_base_key.rs
@@ -1,7 +1,7 @@
 use std::mem::replace;
 
 use crate::{
-    crypto::{cipher_suite::CipherSuite, key_derivation::Ratcheting},
+    crypto::{cipher_suite::CipherSuiteParams, key_derivation::Ratcheting},
     error::Result,
     CipherSuiteVariant,
 };
@@ -15,7 +15,7 @@ use super::ratcheting_key_id::RatchetingKeyId;
 ///
 /// The original key material is not stored for security reasons.
 pub struct RatchetingBaseKey {
-    cipher_suite: CipherSuite,
+    cipher_suite: CipherSuiteParams,
     current_material: Vec<u8>,
     key_id: RatchetingKeyId,
 }